+++ packages/loadout/src/attack/interface.cairo
@@ -4,8 +4,8 @@ use crate::attack::{Attack, AttackWithName, Effect};
 pub trait IAttack<TContractState> {
     fn attack(self: @TContractState, id: felt252) -> Attack;
     fn attacks(self: @TContractState, ids: Array<felt252>) -> Array<Attack>;
-    fn speed(self: @TContractState, id: felt252) -> u8;
-    fn speeds(self: @TContractState, ids: Array<felt252>) -> Array<u8>;
+    fn speed(self: @TContractState, id: felt252) -> u32;
+    fn speeds(self: @TContractState, ids: Array<felt252>) -> Array<u32>;
     fn accuracy(self: @TContractState, id: felt252) -> u8;
     fn accuracies(self: @TContractState, ids: Array<felt252>) -> Array<u8>;
     fn cooldown(self: @TContractState, id: felt252) -> u8;
@@ -17,7 +17,7 @@ pub trait IAttack<TContractState> {
     fn attack_id(
         self: @TContractState,
         name: ByteArray,
-        speed: u8,
+        speed: u32,
         accuracy: u8,
         cooldown: u8,
         hit: Array<Effect>,
@@ -32,7 +32,7 @@ pub trait IAttackAdmin<TContractState> {
     fn create_attack(
         ref self: TContractState,
         name: ByteArray,
-        speed: u8,
+        speed: u32,
         accuracy: u8,
         cooldown: u8,
         hit: Array<Effect>,
